Created Small and Large Explosions
I created the small and large explosions using the particle system. I
imported two materials and two textures into the project. Both the
materials and textures are in a folder called "SimpleParticlePack" As
for the explosions, I created them and made two prefabs for them. The
two prefabs for the explosions are located in the Prefab folder.