credits
requirements:
Update startmenu script
CreditScreen prefab
CreditsSript
CreditsSceenController
Credits.anim

Instructions:
Open on the Menu sceen
Bring in the new CreditScreen prefab
and Insert it as a child of GUI
assign it the CreditScript
Drag Main Menu into the designate slot of that script
Now take the creditscript game object (that you just made) and insert it
into the designated slot of Main Menu game object.
Afterwards, make sure the CreditScreen gameobject has the controller,
and the Credits.anim inserted intot hat controller.

Finally, create a new button and call it "credits"
In the event system for this button call the StartMenscript.Credits()

That should be all. Any questions let me know

-Chris