added main menu theme to main menu and credits theme
hey guys and gals, I added the main menu theme to both the main menu and
credits panel. I added an audio source to the main camera and added the
song to the camera. I changed the volume from 1 to 0.6 so the sound will
not be blasting through the speakers.