Added Main Menu Theme to both the Main Menu and Credits Screen
Hey everybody, I added the Main Menu Music to the Main Menu and Credits
Screen. I also added a folder called  Background Music. This folder will
be located in the Sound folder and will host all of the background
music.  I also did one small change on the title of our game located in
the top panal. I just centered out the text so it would look better in
the scene.