I added the Tank Battle Theme to the Game
Hey everybody, I added the Battle theme to the game. Its located in the
UnetStrikesBack Scene in the Canvas. I added a Audio Source and added
the song in there. I also checkmarked the Box so the song will loop as
well. Ive tested it a few times and It works.