tank now spawns upright when flipped
the flip time is a public variable in the movement script.

I also made it so players cannot drive when upside down.

All the build master has to do is update the playermovement script. I
added a courotine, an isFlipped bool, and flipTimer int, and additional
information in fixed update. That is all.