Added The Lobby Theme Music to Game
Hey everybody, I added the lobby theme music to the game. I added a
function to the play button that will stop the main theme music and the
lobby music to play. I also added a lobby music prefab that contains the
lobby music game object. The empty game object has an audio source with
the lobby music attached to it. Its an object. So when it loads the main
game screen, The battle theme can play. Also I had to re-implement the
battle theme. I added the battle theme music to the scene object with
all of the scene objects.