Added Lobby Music and Battle Music
Hey everybody, I added both the lobby music and the battle music to the
game. The lobby music is located in a prefab i made called
LobbyThemeMusic. Its in there so that when the game scene is called, the
lobby music will actually stop instead of continue to play when the game
is going on. When the play button is pushed, the Main Menu theme stops
playing and the Lobby Music starts. As for the Battle Music, I added
that to the canvas thats in the UNetStrikesBack Scene.